tigerapps/requirements.txt
Django==1.3.1
mysql-python==1.2.3
python-memcached==1.48
django-paypal==0.1.2
django-html-sanitizer==0.1.4
python-dateutil==1.5
twilio==3.3.6
djangorestframework==2.3.14
rest_framework_ember==1.0.3
